backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch>=1.11.0
transformers>=4.21.0
numpy>=1.21.0
python-multipart>=0.0.6
requests>=2.25.0
sentence-transformers>=2.2.0
pydantic>=1.10.0
PyJWT>=2.4.0